show Gantt chart link on simple search when we already searched
The previous link to the Gantt chart on simple search added in 01ab4582
added the link not only on the search results page, but also on the
simple search page if you navigated there without a search. Since
there's nothing to graph yet, it doesn't make sense at that point, and
so we update to show only when there is a query.